🔧 ZS8AE7 code: split branch naming coverage
